requirements.txt
Django>=5.0
djangorestframework>=3.14.0
djangorestframework-simplejwt>=5.3.0
cryptography>=42.0
requests>=2.28.0
openpyxl>=3.1.0
reportlab>=4.0.0
qrcode[pil]>=7.0.0
python-quickbooks>=0.9.0
celery>=5.3.0
redis>=5.0.0
django-celery-beat>=2.5.0
channels>=4.0.0
channels-redis>=4.2.0
daphne>=4.0.0
gunicorn>=22.0.0
dj-database-url>=2.0.0